Moved tag methods, changed tests.
Tag methods were moved to a separate file for better organization. Will likely do this for files too and change the function names to simply be , etc. Considering reimplementing an ORM but resisting that in favor of a purely functional approach.

Added a verbosity switch for the test script.

Rewrote some tests to increase the coverage a bit. Had to fix some bugs that cropped up as a result.