Tests passing.
I need to refactor my code at this point. The mess of functions has tripped me up -  is distinct from the straight  function in tags, etc. I still want to do purely functional so I'm thinking of using modules-as-classes here.

The role of api.py is uncertain. I think I'll use it as a converter of sorts - each module will have a distinct way of entering input and api.py wrangles it neatly. Important for tags and filepaths.