Some minor tweaks.
I'm stumped on the parsing part. I'll have to revisit the original code to add in that modularity at the root level.

It's the compound condition part: I need to be able to translate a string into SQL safely. I can see the merits of SQLAlchemy Core much  more clearly.